Refactor image processing out of board_canvas.py
